create-react-app: mgirate login
