--- app/components/login/index.js
@@ -1,1 +1,0 @@
-export { default } from './scene.jsx'
