--- app/components/login/login.jsx
@@ -1,61 +1,0 @@
-import React from 'react'
-import { connect } from 'react-redux'
-import { routeActions } from 'react-router-redux'
-import { Panel } from 'react-bootstrap'
-
-import { loginUser, addFlashMessage } from 'actions'
-
-import Form from './form.jsx'
-
-class Login extends React.Component {
-  constructor(props) {
-    super(props)
-    this.handleSubmit = this.handleSubmit.bind(this)
-    this.afterLogin   = this.afterLogin.bind(this)
-  }
-
-  componentDidMount() {
-    if (this.props.isAuthorized) {
-      this.afterLogin()
-    }
-  }
-
-  handleSubmit(values) {
-    const { login } = this.props
-    const { email, password } = values
-
-    return login(email, password)
-  }
-
-  afterLogin() {
-    this.props.addFlashMessage('You successfully signed in.')
-    this.props.redirectToRoot()
-  }
-
-  render() {
-    return(
-      <Panel>
-        <Form onSubmit={ this.handleSubmit } onSubmitSuccess={ this.afterLogin } />
-      </Panel>
-    )
-  }
-}
-
-Login.propTypes = {
-  isAuthorized:    React.PropTypes.bool,
-  login:           React.PropTypes.func.isRequired,
-  redirectToRoot:  React.PropTypes.func.isRequired,
-  addFlashMessage: React.PropTypes.func.isRequired,
-}
-
-const select = (state) => ({
-  isAuthorized: !!state.auth.token,
-})
-
-const dispatcher = (dispatch) => ({
-  login:           (email, password) => new Promise((res, rej) => dispatch(loginUser(email, password, res, rej))),
-  redirectToRoot:  () => dispatch(routeActions.push('/')),
-  addFlashMessage: (message, type = null) => dispatch(addFlashMessage(message, type)),
-})
-
-export default connect(select, dispatcher)(Login)
